chore(node): move the selection to the cloud service manager
