+++ node/src/main/java/eu/cloudnetservice/node/service/defaults/NodeCloudServiceFactory.java
@@ -16,7 +16,6 @@
 
 package eu.cloudnetservice.node.service.defaults;
 
-import com.google.common.collect.ComparisonChain;
 import eu.cloudnetservice.driver.channel.ChannelMessage;
 import eu.cloudnetservice.driver.channel.ChannelMessageTarget;
 import eu.cloudnetservice.driver.event.EventManager;
@@ -27,7 +26,6 @@ import eu.cloudnetservice.driver.service.GroupConfiguration;
 import eu.cloudnetservice.driver.service.ServiceConfiguration;
 import eu.cloudnetservice.driver.service.ServiceInfoSnapshot;
 import eu.cloudnetservice.node.Node;
-import eu.cloudnetservice.node.cluster.NodeServer;
 import eu.cloudnetservice.node.cluster.NodeServerProvider;
 import eu.cloudnetservice.node.event.service.CloudServiceNodeSelectEvent;
 import eu.cloudnetservice.node.network.listener.message.ServiceChannelMessageListener;
@@ -84,7 +82,7 @@ public class NodeCloudServiceFactory implements CloudServiceFactory {
         var nodeServer = nodeSelectEvent.nodeServer();
         if (nodeServer == null) {
           // no node was set by the event, try to select a node or return if no node can pick up the service
-          nodeServer = this.selectNodeForService(serviceConfiguration);
+          nodeServer = this.serviceManager.selectNodeForService(serviceConfiguration);
           if (nodeServer == null) {
             return null;
           }
@@ -120,50 +118,6 @@ public class NodeCloudServiceFactory implements CloudServiceFactory {
     }
   }
 
-  public @Nullable NodeServer selectNodeForService(@NonNull ServiceConfiguration configuration) {
-    // check if the node is already specified
-    if (configuration.serviceId().nodeUniqueId() != null) {
-      // check for a cluster node server
-      var server = this.nodeServerProvider.node(configuration.serviceId().nodeUniqueId());
-      if (server != null) {
-        // the requested node is a cluster node, check if that node is still accepting services
-        return !server.available() || server.nodeInfoSnapshot().draining() ? null : server;
-      }
-      // no node server with the given name which can start services found
-      return null;
-    }
-
-    // find the best node server
-    return this.nodeServerProvider.nodeServers().stream()
-      .filter(NodeServer::available)
-      .filter(nodeServer -> !nodeServer.nodeInfoSnapshot().draining())
-      .filter(server -> {
-        var allowedNodes = configuration.serviceId().allowedNodes();
-        return allowedNodes.isEmpty() || allowedNodes.contains(server.info().uniqueId());
-      })
-      .min((left, right) -> {
-        // calculate the reserved memory amount based on the cached service information on this node
-        // this is the better way to do this, as newly created services on other nodes will get cached instantly, rather
-        // than us needing to wait for the updated node info to be sent by the associated node. In normal scenarios
-        // that is not a big problem, however when many start requests are coming in, that can lead to one node picking
-        // up a lot of services until (only a few ms later) the updated snapshot is present.
-        var leftReservedMemory = this.calculateReservedMemoryPercentage(left);
-        var rightReservedMemory = this.calculateReservedMemoryPercentage(right);
-
-        // we elevate the used heap memory percentage over the cpu usage, as it's varying much more
-        var chain = ComparisonChain.start().compare(leftReservedMemory, rightReservedMemory);
-        // only include the cpu usage if both nodes can provide a value
-        if (left.nodeInfoSnapshot().processSnapshot().systemCpuUsage() >= 0
-          && right.nodeInfoSnapshot().processSnapshot().systemCpuUsage() >= 0) {
-          // add the system usage to the chain
-          chain = chain.compare(
-            left.nodeInfoSnapshot().processSnapshot().systemCpuUsage(),
-            right.nodeInfoSnapshot().processSnapshot().systemCpuUsage());
-        }
-        // use the result of the comparison
-        return chain.result();
-      }).orElse(null);
-  }
 
   protected @Nullable ServiceInfoSnapshot sendNodeServerStartRequest(
     @NonNull String message,
@@ -245,15 +199,5 @@ public class NodeCloudServiceFactory implements CloudServiceFactory {
     }
     // set the new unique id
     output.uniqueId(uniqueId);
-  }
-
-  protected int calculateReservedMemoryPercentage(@NonNull NodeServer server) {
-    // get the reserved memory on the given node based on the services which are running on it and sum it up
-    var reservedMemory = this.serviceManager.services().stream()
-      .filter(info -> info.serviceId().nodeUniqueId().equals(server.name()))
-      .mapToInt(info -> info.configuration().processConfig().maxHeapMemorySize())
-      .sum();
-    // convert to a percentage
-    return (reservedMemory * 100) / server.nodeInfoSnapshot().maxMemory();
   }
 }
